+++ NEWS.md
@@ -1,8 +1,11 @@
-# cancensus - 0.5.1
+# cancensus - 0.5.2
+
 ## Major changes
 - Added functionality for users to detect and remove locally cached data that has been recalled by Statistics Canada
 - added a check on initial `get_census` call that produces a warning if locally cached data has been recalled by Statistics Canada
 - added option to retrieve higher-resolution geometries in `get_census`
+
+# cancensus - 0.5.1
 
 ## Minor changes
 - renamed get/set cache/api_key functions to avoid name conflicts with related packages